ai_service/requirements.txt
fastapi
uvicorn
python-dotenv
gitpython
faiss-cpu
requests
numpy
supabase
scikit-learn
pydantic